+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-03 22:48:24
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-08-04 11:25:57
+// Commit: perf(core): optimize memory consumption in core list viewport
 
 class coreComponentState {
-  final String commitRef = '3335';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-08-03T22:48:24+05:00');
+  final String commitRef = '3341';
+  final String lastMessage = 'perf(core): optimize memory consumption in core list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-04T11:25:57+05:00');
 }